+++ src/apps/client/report/report_crud.h
@@ -52,6 +52,7 @@ namespace tictac_track {
       bool Reset();
 
       bool IsAnyEntryRunning();
+      bool CurrentDayHasTasks();
 
     private:
       enum EntryStatus {